+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.3] - 2018-11-08
+## [1.1.0] - 2018-11-08
+### Added
+- Support for increasing quantity on non subscription products on the cart page
+- Cache for subscription products
 ### Fixed
 - Race condition for adding products on PDP